perupdatean tod

+++ SkyFoxRobot/modules/sql/global_bans_sql.py
@@ -1,12 +1,12 @@
 import threading
+
 from SkyFoxRobot.modules.sql import BASE, SESSION
-from sqlalchemy import Boolean, Column, String, UnicodeText
-from sqlalchemy.sql.sqltypes import BigInteger
+from sqlalchemy import Boolean, Column, Integer, String, UnicodeText
 
 
 class GloballyBannedUsers(BASE):
     __tablename__ = "gbans"
-    user_id = Column(BigInteger, primary_key=True)
+    user_id = Column(Integer, primary_key=True)
     name = Column(UnicodeText, nullable=False)
     reason = Column(UnicodeText)
 
@@ -161,6 +161,6 @@ def migrate_chat(old_chat_id, new_chat_id):
         SESSION.commit()
 
 
-# create in memory userid to avoid disk access
+# Create in memory userid to avoid disk access
 __load_gbanned_userid_list()
 __load_gban_stat_list()
